% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.r
\name{exec}
\alias{exec}
\title{Execute a GET request with the given URL and query params}
\usage{
exec(path, params = list(), as = "parsed", token = get_access_cred())
}
\description{
Execute a GET request with the given URL and query params
}
\section{Fields}{

\describe{
\item{\code{path}}{The URL to GET}

\item{\code{params}}{The URL parameters}

\item{\code{as}}{The type of the response}
}}

