% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.r
\docType{class}
\name{gcs_store-class}
\alias{gcs_store}
\alias{gcs_store-class}
\title{A reference class to abstract a Gcs Store}
\description{
A reference class to abstract a Gcs Store
}
\section{Fields}{

\describe{
\item{\code{bucket}}{Google Cloud Storage Bucket}
}}
\examples{
store <- gcs_store$new("my_bucket")
files <- store$list_files() 
file <- store$download_file("path/to/my/file.csv")
store$upload_file("path/to/my/file.csv", mtcars)

}

