% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{post}
\alias{post}
\title{Execute a POST request with the given URL and request body}
\usage{
post(url, body, params = list(), token = get_access_cred())
}
\arguments{
\item{params}{The query parameters}
}
\description{
Execute a POST request with the given URL and request body
}
\section{Fields}{

\describe{
\item{\code{url}}{The URL to POST to}

\item{\code{body}}{The json-able Request Body}
}}

