% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgcs.r
\docType{pacakge}
\name{Rgcs}
\alias{Rgcs}
\alias{rgcs}
\title{An R interface to Google Cloud Storage}
\description{
For more information about how Rgcs works, please see the Readme. The most 
important method to get started is \code{\link{list_buckets}}.
}

