% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/store.r
\name{list_buckets}
\alias{list_buckets}
\title{A helper funciton to list the GCS buckets specific to a project}
\usage{
list_buckets(project, max_results = NULL)
}
\description{
A helper funciton to list the GCS buckets specific to a project
}
\section{Fields}{

\describe{
\item{\code{project}}{The GCS project name}

\item{\code{max_results}}{The maximum number of buckets to return}
}}

